refactoring JSON classes into one package
refactoring JSON classes into one package
