+++ testar/resources/settings/desktop_simple_stategraph_sikulix/Protocol_desktop_simple_stategraph_sikulix.java
@@ -33,12 +33,8 @@ import java.io.File;
 import java.util.Set;
 
 import nl.ou.testar.SimpleGuiStateGraph.GuiStateGraphWithVisitedActions;
-import nl.ou.testar.HtmlSequenceReport;
-import nl.ou.testar.RandomActionSelector;
-import org.fruit.Drag;
+import nl.ou.testar.HtmlReporting.HtmlSequenceReport;
 import org.fruit.Util;
-import org.fruit.alayer.AbsolutePosition;
-import org.fruit.alayer.Point;
 import org.fruit.alayer.Action;
 import org.fruit.alayer.exceptions.*;
 import org.fruit.alayer.SUT;
